Consolidate groups of similar functionality
