--- src/backend/dao/userDao.js
@@ -1,63 +1,0 @@
-const {
-  DynamoDBClient,
-  QueryCommand,
-  ScanCommand,
-} = require("@aws-sdk/client-dynamodb");
-const {
-  DynamoDBDocumentClient,
-  GetCommand,
-  PutCommand,
-  UpdateCommand,
-  DeleteCommand,
-} = require("@aws-sdk/lib-dynamodb");
-
-const encounterDao = require("../dao/encounterDao");
-
-const client = new DynamoDBClient({ region: "us-east-1" });
-
-const documentClient = DynamoDBDocumentClient.from(client);
-
-const TableName = "Dungeon_Delver_Users";
-
-const getUserByEmail = async (email) => {
-  try {
-    const command = new GetCommand({
-      TableName,
-      Key: { email },
-    });
-
-    const data = await documentClient.send(command);
-
-    return data.Item || null;
-  } catch (err) {
-    console.error(err);
-    throw { status: 500, message: "Error retrieving  user by email" };
-  }
-};
-
-const getUserByUsername = async (username) => {
-  try {
-    const command = new QueryCommand({
-      TableName,
-      IndexName: "username-index",
-      KeyConditionExpression: "#username = :username",
-      ExpressionAttributeNames: {
-        "#username": "username",
-      },
-      ExpressionAttributeValues: {
-        ":username": { S: username },
-      },
-    });
-
-    const data = await documentClient.send(command);
-
-    return data.Items[0] || null;
-  } catch (err) {
-    throw { status: 500, message: "Error retrieving user by username" };
-  }
-};
-
-module.exports = {
-  getUserByEmail,
-  getUserByUsername,
-};
