+++ src/server.js
@@ -1,13 +1,11 @@
 require("dotenv").config();
 const express = require("express");
 const app = express();
-const userController = require("./backend/controller/userController");
-const accountController = require("./backend/controller/AccountController");
 const { logger } = require("./backend/utils/logger");
+const accountController = require("./backend/controller/AccountController");
 
 app.use(express.json());
 
-app.use("/api/users", userController);
 app.use("/api/account", accountController);
 
 const PORT = process.env.PORT || 3000;
